+++ src/components/Home/HomeHeader/HomeHeader.tsx
@@ -157,8 +157,6 @@ export const HomeHeader = (headerData: IHomeHeader) => {
           </div>
         </div>
       )}
-      {!isPlaying && <img src="/assets/first_frame.jpg" alt="" className="grayscale" />}
-
       <VideoHeader isPlaying={isPlaying} pauseVideo={pauseVideo} ref={videoRef} />
       <AllowCookies />
     </div>
